refactor: Split out RateLimiterLayer into a RequestRateLimiterLayer and a BandwidthRateLimiterLayer. Also cleanup fns a bit.
