+++ pubky-homeserver/src/data_directory/quota_config/mod.rs
@@ -1,17 +1,16 @@
-mod bandwidth_rate;
+mod bandwidth_quota;
 mod glob_pattern;
 mod http_method;
 mod limit_key;
 mod path_limit;
-mod quota_value;
-mod rate_unit;
+pub(crate) mod rate_unit;
+mod request_count_quota;
 mod time_unit;
 
-pub use bandwidth_rate::BandwidthRate;
+pub use bandwidth_quota::BandwidthQuota;
 pub use glob_pattern::GlobPattern;
 pub use http_method::HttpMethod;
 pub use limit_key::{LimitKey, LimitKeyType};
 pub use path_limit::*;
-pub use quota_value::QuotaValue;
-pub use rate_unit::RateUnit;
+pub use request_count_quota::RequestCountQuota;
 pub use time_unit::TimeUnit;
