+++ pubky-homeserver/src/persistence/sql/entities/signup_code.rs
@@ -19,7 +19,7 @@ impl SignupCodeRepository {
     /// Create a new signup code with the given limits for users who redeem it.
     /// The executor can either be db.pool() or a transaction.
     ///
-    /// Rate limit strings are validated by roundtripping through `BandwidthRate`
+    /// Rate limit strings are validated by roundtripping through `BandwidthQuota`
     /// parsing to ensure only well-formed values reach the database.
     pub async fn create<'a>(
         id: &SignupCodeId,
@@ -329,7 +329,7 @@ mod tests {
     #[tokio::test]
     #[pubky_test_utils::test]
     async fn test_create_with_quota() {
-        use crate::data_directory::quota_config::BandwidthRate;
+        use crate::data_directory::quota_config::BandwidthQuota;
         use crate::shared::user_quota::QuotaOverride;
         use std::str::FromStr;
 
@@ -338,7 +338,7 @@ mod tests {
 
         let config = UserQuota {
             storage_quota_mb: QuotaOverride::Value(500),
-            rate_read: QuotaOverride::Value(BandwidthRate::from_str("100mb/m").unwrap()),
+            rate_read: QuotaOverride::Value(BandwidthQuota::from_str("100mb/m").unwrap()),
             rate_write: QuotaOverride::Default,
             ..Default::default()
         };
@@ -446,15 +446,15 @@ mod tests {
     #[tokio::test]
     #[pubky_test_utils::test]
     async fn test_signup_token_limits_applied_to_user() {
-        use crate::data_directory::quota_config::BandwidthRate;
+        use crate::data_directory::quota_config::BandwidthQuota;
         use crate::persistence::sql::user::UserRepository;
         use crate::shared::user_quota::QuotaOverride;
         use std::str::FromStr;
 
         let db = SqlDb::test().await;
 
-        fn bw(s: &str) -> BandwidthRate {
-            BandwidthRate::from_str(s).unwrap()
+        fn bw(s: &str) -> BandwidthQuota {
+            BandwidthQuota::from_str(s).unwrap()
         }
 
         // 1) Create a signup code with custom limits
